Adds python script and github action for testing Pysa
Adds python script which compares the issues pysa detects in
delibrately_vulnerable_app and raise exceptions (return an error) when
there is a mismatch. The output is also written to `results.actual` and
printed.

Adds a github action called pysa that install the latest nightly version
of pyre-check, runs setup.sh, and the run_integration_tests.py python
script. It thereby, effectively validating the Pysa models.

Adds results.expected that will contain issues known to be true in the
app. For now, it's just issues from rce.py, but can be expanded later
on as more vulnerabilities are added to the flask app. The issues
contained in it is in a concise form containing only the attributes-
path, define, and code.

The python script, run_integration_tests.py reads issues in
resulsts.expected, runs pysa, compares issues generated by pysa
against those specified manually in results.expected and prints a list
of issues which were detected in addition to the specified ones and/or
the issues that weren't detected but were in resulsts.expected, in
addition to raising a custom exception: PysaIssuesMismatchException.

Adds .gitignore containing list of directories generated by running pysa
via the command `pyre analyse --no-verify`.

Modifies setup.sh to tailor to the github actions by creating a
.pyre_configuration file with tain_models_path pointing to ../../stubs

Adds requirements.txt to contain requirements for the flask app.
Currently it contains only `flask` but will be expanded later when more
vulnerabilities are added to flask app.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>